@@ -9,33 +9,35 @@
     "test": "mocha test/*.coffee"
   },
   "dependencies": {
+    "@mdi/js": "^5.0.45",
+    "@vue/cli": "^4.1.1",
     "c3": "^0.7.4",
+    "global": "^4.4.0",
     "jsOAuth2": "https://github.com/twhtanghk/jsOAuth2",
-    "vue": "^2.5.17",
+    "model": "https://github.com/twhtanghk/model",
+    "register-service-worker": "^1.6.2",
+    "vue": "^2.6.11",
     "vuelidate": "^0.7.4"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.9.2",
-    "@vue/cli-plugin-eslint": "^3.9.2",
-    "@vue/cli-service": "^3.9.3",
+    "@vue/cli-plugin-babel": "^4.1.1",
+    "@vue/cli-plugin-eslint": "^4.1.1",
+    "@vue/cli-plugin-pwa": "~4.2.0",
+    "@vue/cli-service": "^4.1.1",
     "babel-eslint": "^10.0.2",
-    "bootstrap-vue": "^2.0.0-rc.27",
     "coffee-loader": "^0.9.0",
     "coffeescript": "^2.3.2",
-    "compression-webpack-plugin": "^3.0.0",
+    "compression-webpack-plugin": "^3.0.1",
     "dotenv": "^8.0.0",
-    "eslint": "^6.1.0",
-    "eslint-plugin-vue": "^5.2.3",
-    "material-icons": "^0.3.0",
+    "eslint": "^6.8.0",
+    "eslint-plugin-vue": "^6.0.1",
     "mocha": "^6.2.0",
     "node-sass": "^4.12.0",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^8.0.0",
     "supertest": "^4.0.2",
     "vue-router": "^3.0.7",
-    "vue-template-compiler": "^2.5.17",
-    "vue.model": "https://github.com/twhtanghk/vue.model",
-    "vue.oauth2": "https://github.com/twhtanghk/vue.oauth2",
-    "vuetify": "^2.0.5"
+    "vue-template-compiler": "^2.6.11",
+    "vuetify": "^2.1.15"
   },
   "eslintConfig": {
     "root": true,
