@@ -1,16 +1,16 @@
 <template>
-  <v-layout row wrap>
+  <v-row wrap justify='center'>
     <ap/>
     <sta/>
     <card header='Settings'>
       <v-btn color='primary' @click='factory'>Factory Default</v-btn>
       <v-btn color='secondary' @click='reboot'>Reboot</v-btn>
     </card>
-  </v-layout>
+  </v-row>
 </template>
 
 <script lang='coffee'>
-{cfg} = require('./model').default
+{Cfg} = require('./plugins/model.coffee').default
 
 export default
   components:
@@ -19,11 +19,11 @@ export default
     card: require('./card').default
   methods:
     factory: ->
-      cfg
-        .get url: "#{cfg.baseUrl}/factory"
+      Cfg
+        .get url: "#{Cfg.baseUrl}/factory"
         .catch console.error
     reboot: ->
-      cfg
-        .get url: "#{cfg.baseUrl}/reboot"
+      Cfg
+        .get url: "#{Cfg.baseUrl}/reboot"
         .catch console.error
 </script>
